feature: move the Datasource object inside the data package
Coherently with the Android implementation best practices, put the data sources inside a package that will be referred as repository. In this case, the Datasource object plays the repository role, and it builds necessary data by exploiting the content in the resources.

Integrates #9
