+++ app/src/main/java/com/example/mycity/ui/Categories.kt
@@ -23,8 +23,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.mycity.Datasource
-import com.example.mycity.R
+import com.example.mycity.data.Datasource
 import com.example.mycity.model.Category
 import com.example.mycity.ui.theme.MyCityTheme
